Improvements for Condition module
This implements a generic method get() in the Request object
to retrieve request attributes or session variables.

It support the following syntax for retrieving variables:

To retrieve request attributes:

  request:<variable>
  r:<variable>

If no scope is given it defaults to 'request'

  <variable>

To retrieve session variables:

  session:<variable>
  s:<variable>

It also changes the Condition plugin to accept this syntax in
the "key" configuration parameter.

Also an `invert` option has been added to the condition module.